First round of code review and added unit tests
